// api/models/CollectionListItemBatchAddRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 批量添加收藏到列表请求
 */
export type CollectionListItemBatchAddRequest = {
    /**
     * 收藏列表ID
     */
    collectionListId: number;
    /**
     * 收藏ID列表
     */
    collectionIds: Array<number>;
};