// api/models/CollectionTagAddRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 添加收藏标签请求
 */
export type CollectionTagAddRequest = {
    /**
     * 收藏ID
     */
    collectionId: number;
    /**
     * 标签名称
     */
    tagName: string;
};