// api/models/CollectionTagDeleteRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 删除收藏标签请求
 */
export type CollectionTagDeleteRequest = {
    /**
     * 收藏ID
     */
    collectionId: number;
    /**
     * 标签名称
     */
    tagName: string;
};